Add files via upload
This is task two that I performed during my codsoft internship.This program is a simple yet practical tool for calculating and displaying a student's performance based on the marks they have obtained in different subjects. It takes the user's input, processes it, and provides clear results, including the total marks, average percentage, and corresponding grade.